@@ -12,21 +12,38 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
-    "@heroui/react": "^2.7.2",
-    "@tailwindcss/forms": "^0.5.10",
-    "next": "^15.1.7",
-    "next-auth": "^4.24.11",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "swiper": "^11.2.4"
-  },
+  "@heroui/button": "2.2.13",
+  "@heroui/code": "2.2.10",
+  "@heroui/input": "2.4.13",
+  "@heroui/kbd": "2.2.10",
+  "@heroui/link": "2.2.11",
+  "@heroui/listbox": "2.3.13",
+  "@heroui/navbar": "2.2.12",
+  "@heroui/snippet": "2.2.14",
+  "@heroui/switch": "2.2.12",
+  "@heroui/system": "2.4.10",
+  "@heroui/theme": "2.4.9",
+  "@heroui/react": "^2.7.4",
+  "@tailwindcss/forms": "^0.5.10",
+  "motion": "^12.4.7",
+  "next": "15.2.0",
+  "next-auth": "^4.24.11",
+  "react": "^19.0.0",
+  "react-dom": "^19.0.0",
+  "swiper": "^11.2.4",
+  "framer-motion": "^12.4.7"
+},
+
   "devDependencies": {
-    "@tailwindcss/postcss": "^4.0.9",
-    "@tailwindcss/postcss7-compat": "^2.2.17",
+    "@eslint/eslintrc": "^3",
+    "@tailwindcss/cli": "^4.0.9",
+    "@tailwindcss/forms": "^0.5.10",
+    "@tailwindcss/postcss": "^4",
+    "@tailwindcss/typography": "^0.5.16",
     "autoprefixer": "^10.4.20",
+    "eslint": "^9",
+    "eslint-config-next": "15.2.0",
     "postcss": "^8.5.3",
-    "postcss-import": "^16.1.0",
-    "postcss-nesting": "^13.0.1",
-    "tailwindcss": "^4.0.8"
+    "tailwindcss": "^3.4.17"
   }
 }
