@@ -1,14 +1,47 @@
 import NextAuth from "next-auth";
+import CredentialsProvider from "next-auth/providers/credentials";
 import GoogleProvider from "next-auth/providers/google";
 
 export default NextAuth({
   providers: [
+    CredentialsProvider({
+      name: "Credentials",
+      credentials: {
+        email: { label: "Email", type: "text" },
+        password: { label: "Password", type: "password" },
+        role: { label: "Role", type: "text" },
+      },
+      async authorize(credentials) {
+        const { email, password, role } = credentials;
+        
+        // Mock database check (Replace with actual DB check)
+        if (email === "client@example.com" && password === "password" && role === "client") {
+          return { id: 1, name: "Client User", email, role };
+        }
+        if (email === "business@example.com" && password === "password" && role === "business") {
+          return { id: 2, name: "Business User", email, role };
+        }
+        return null; // Fail authentication
+      },
+    }),
     GoogleProvider({
       clientId: process.env.GOOGLE_CLIENT_ID,
       clientSecret: process.env.GOOGLE_CLIENT_SECRET,
     }),
   ],
+  callbacks: {
+    async session({ session, token }) {
+      session.user.role = token.role;
+      return session;
+    },
+    async jwt({ token, user }) {
+      if (user) {
+        token.role = user.role;
+      }
+      return token;
+    },
+  },
   pages: {
-    signIn: "/login",  // Redirect to login page if not authenticated
+    signIn: "/service/login",
   },
 });
