@@ -1,8 +1,6 @@
 module.exports = {
   plugins: {
-    'postcss-import': {},
-    'postcss-nesting': {},
-    '@tailwindcss/postcss': {},  // Change this line from 'tailwindcss': {}
-    'autoprefixer': {},
+    tailwindcss: {},
+    autoprefixer: {},
   },
 }
